dependency injection bootstrapper + reorganization
